mongo_vs_postgres/requirements.txt
pydantic-settings==2.0.3
clickhouse-driver==0.2.6
pymongo==4.5.0
psycopg2-binary==2.9.7
Faker==19.6.2